# openapi.yaml
openapi: 3.0.3
info:
  title: Bhesi Content API
  version: 1.0.0
  description: REST API for admin authentication, blogs, and case studies.
servers:
  - url: http://localhost:4000
paths:
  /api/auth/login:
    post:
      summary: Login admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 6 }
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/auth/refresh:
    post:
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '200':
          description: New tokens
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: object
                    properties:
                      token: { type: string }
                      refreshToken: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /api/auth/me:
    get:
      security: [ bearerAuth: [] ]
      summary: Current authenticated admin
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { $ref: '#/components/schemas/AdminUser' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AdminUser:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
...
    AuthResponse:
      type: object
      properties:
        success: { type: boolean }
        data:
          type: object
          properties:
            token: { type: string }
            refreshToken: { type: string }
            user: { $ref: '#/components/schemas/AdminUser' }
  responses:
    Unauthorized:
      description: Unauthorized
...